# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/company/advt/upload": {
            "put": {
                "description": "Company upload advertisement",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Company upload advertisement",
                "parameters": [
                    {
                        "description": "name: advertisementInfo",
                        "name": "advertisementInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CompanyUploadAdvtReq"
                        }
                    },
                    {
                        "description": "name: image",
                        "name": "image",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/company/login": {
            "post": {
                "description": "Company Login by account and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Company login",
                "parameters": [
                    {
                        "description": "company login request",
                        "name": "request_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CompanyLoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company login response",
                        "schema": {
                            "$ref": "#/definitions/response.CompanyLoginResp"
                        }
                    }
                }
            }
        },
...
        "/company/register": {
            "post": {
                "description": "Company register",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Company register",
                "parameters": [
                    {
                        "description": "company register request",
                        "name": "request_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CompanyRegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response"
                    }
                }
            }
        },
...
        "/company/update-info": {
            "post": {
                "description": "Company update info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Company update info",
                "parameters": [
                    {
                        "description": "company update info  request",
                        "name": "request_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CompanyUpdateInfoReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company UpdateInfoReq response"
                    }
                }
            }
        },
        "/company/update-pwd": {
            "post": {
                "description": "Company update password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Company update password",
                "parameters": [
                    {
                        "description": "company update password request",
                        "name": "request_body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.CompanyUpdatePwdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company update password response"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "advertisement_model.AdvertisementInfo": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "2023-06-04"
                },
                "jump_to_url": {
                    "type": "string",
                    "example": "example.com/1919810.jpg"
                },
                "position": {
                    "type": "integer",
                    "example": 1
                },
                "start_date": {
                    "type": "string",
                    "example": "2023-05-14"
                },
                "title": {
                    "type": "string",
                    "example": "advertisement example"
                }
            }
        },
...
        "request.CompanyLoginReq": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "request.CompanyRegisterReq": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string",
                    "example": "联创"
                },
                "address": {
                    "type": "string",
                    "example": "启明学院亮胜楼八楼"
                },
                "business_license_number": {
                    "type": "string",
                    "example": "114514"
                },
                "manager_name": {
                    "type": "string",
                    "example": "汉堡"
                },
                "manager_tel": {
                    "type": "string",
                    "example": "1919810"
                },
                "name": {
                    "type": "string",
                    "example": "联小创在线科技有限公司"
                },
                "password": {
                    "type": "string",
                    "example": "uniquestudio"
                }
            }
        },
        "request.CompanyUpdateInfoReq": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string",
                    "example": "联创"
                },
                "address": {
                    "type": "string",
                    "example": "启明学院亮胜楼八楼"
                },
                "business_license_number": {
                    "type": "string",
                    "example": "114514"
                },
                "manager_name": {
                    "type": "string",
                    "example": "汉堡"
                },
                "manager_tel": {
                    "type": "string",
                    "example": "1919810"
                },
                "name": {
                    "type": "string",
                    "example": "联小创在线科技有限公司"
                }
            }
        },
        "request.CompanyUpdatePwdReq": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "1919810"
                }
            }
        },
        "request.CompanyUploadAdvtReq": {
            "type": "object",
            "properties": {
                "advertisementInfo": {
                    "$ref": "#/definitions/advertisement_model.AdvertisementInfo"
                }
            }
        },
...
        "response.CompanyLoginResp": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string",
                    "example": "联创"
                },
                "token": {
                    "type": "string",
                    "example": "114514"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}